requirements.txt
bson==0.5.8
dataclasses==0.7
Django==2.2.10
django-object-actions==2.0.0
djongo==1.3.0
grpcio==1.27.1
grpcio-tools==1.27.1
mongolog==0.1.2
pkg-resources==0.0.0
protobuf==3.11.3
pymongo==3.10.1
python-dateutil==2.8.1
python-http-client==3.2.4
pytz==2019.3
sendgrid==6.1.1
singlemodeladmin==0.9
six==1.14.0
sqlparse==0.2.4
